+++ QuantLibAddin/qla/objects/dividendvanillaoption.hpp
@@ -31,7 +31,7 @@ namespace QuantLibAddin {
             const std::vector<double>& dividends,
             const std::string &optionTypeID,
             const std::string &payoffID,
-            const float &strike,
+            const double &strike,
             const std::string &exerciseID,
             const long &exerciseDate,
             const long &settlementDate,
