+++ QuantLibAddin/qla/objects/vanillaoption.hpp
@@ -29,7 +29,7 @@ namespace QuantLibAddin {
             const boost::shared_ptr<StochasticProcess> &stochasticProcess,
             const std::string &optionTypeID,
             const std::string &payoffID,
-            const float &strike,
+            const double &strike,
             const std::string &exerciseID,
             const long &exerciseDate,
             const long &settlementDate,
